pyproject.toml
# Copyright 2025 Vijaykumar Singh <singhvjd@gmail.com>
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "victor"
version = "0.1.0"
description = "Enterprise-Ready AI Coding Assistant - Code to Victory with Any AI"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Vijaykumar Singh", email = "singhvjd@gmail.com"}
]
requires-python = ">=3.10"
keywords = [
    "ai",
    "llm",
    "coding-agent",
    "victor",
    "anthropic",
    "openai",
    "ollama",
    "claude",
    "gpt",
    "cli",
    "terminal"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",

    # Async
    "httpx>=0.27",
    "aiofiles>=23.0",

    # CLI/UI
    "typer>=0.12",
    "rich>=13.7",
    "prompt-toolkit>=3.0",

    # LLM Providers
    "anthropic>=0.34",
    "openai>=1.40",
    "google-generativeai>=0.7",

    # Tools
    "gitpython>=3.1",
    "jsonschema>=4.20",
    "tiktoken>=0.7",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",

    # Code Analysis
    "tree-sitter>=0.21",
    "pygments>=2.17",

    # Caching
    "cachetools>=5.3",
    "diskcache>=5.6",
    "docker>=7.0",
    "docstring-parser>=0.15",

    # Semantic Tool Selection (Default: local embeddings)
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "pytest-json-report>=1.5.0",
    "respx>=0.21",

    # Linting/Formatting
    "ruff>=0.5",
    "mypy>=1.10",
    "black>=24.0",

    # Type stubs
    "types-pyyaml",
    "types-aiofiles",
]

docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]

all = [
    "victor[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/vijayksingh/victor"
Documentation = "https://github.com/vijayksingh/victor#readme"
Repository = "https://github.com/vijayksingh/victor"
Issues = "https://github.com/vijayksingh/victor/issues"

[project.scripts]
victor = "victor.ui.cli:app"
vic = "victor.ui.cli:app"

[tool.setuptools.packages.find]
include = ["victor*"]
exclude = ["tests*", "docs*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=victor",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (require external services)",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: marks tests as unit tests",
]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [
    "archive",
    "examples",
    "docker/demos",
    "victor_test",
    "demo_workspace",
    "src",
    "ui",
    "web/ui",
    "htmlcov",
    "venv",
    "game_app.py",
    "tests",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # allow broad HTTPException raising for now
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/**/*.py" = ["B018"]  # useless expression (for test assertions)

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = false
ignore_missing_imports = true
exclude = ""
files = ["victor", "web"]

[[tool.mypy.overrides]]
module = [
    "victor.providers.*",
    "victor.codebase.embeddings.*",
    "victor.tools.*",
    "victor.mcp.*",
]
ignore_errors = true

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.coverage.run]
source = ["victor"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    # CLI modules require interactive terminal testing
    "victor/ui/cli.py",
    "victor/ui/commands.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
